Add BaseConfigLoader and refactor loaders
